+++ cmd/mimir/main_test.go
@@ -6,13 +6,10 @@
 package main
 
 import (
-	"bytes"
 	"flag"
-	"io"
 	"os"
 	"path/filepath"
 	"strings"
-	"sync"
 	"testing"
 
 	"github.com/go-kit/log"
@@ -23,6 +20,7 @@ import (
 
 	"github.com/grafana/mimir/pkg/mimir"
 	"github.com/grafana/mimir/pkg/util/fieldcategory"
+	"github.com/grafana/mimir/pkg/util/test"
 )
 
 func TestFlagParsing(t *testing.T) {
@@ -244,7 +242,7 @@ ruler_storage:
 	} {
 		t.Run(name, func(t *testing.T) {
 			_ = os.Setenv("TARGET", "ingester")
-			testSingle(t, tc.arguments, tc.yaml, []byte(tc.stdoutMessage), []byte(tc.stderrMessage), []byte(tc.stdoutExcluded), []byte(tc.stderrExcluded), tc.assertConfig)
+			testSingle(t, tc.arguments, tc.yaml, tc.stdoutMessage, tc.stderrMessage, tc.stdoutExcluded, tc.stderrExcluded, tc.assertConfig)
 		})
 	}
 }
@@ -284,7 +282,7 @@ func TestHelp(t *testing.T) {
 			t.Cleanup(restoreIfNeeded)
 
 			testMode = true
-			co := captureOutput(t)
+			co := test.CaptureOutput(t)
 
 			const cmd = "./cmd/mimir/mimir"
 			os.Args = []string{cmd, tc.arg}
@@ -307,7 +305,7 @@ func TestHelp(t *testing.T) {
 	}
 }
 
-func testSingle(t *testing.T, arguments []string, configYAML string, stdoutMessage, stderrMessage, stdoutExcluded, stderrExcluded []byte, assertConfig func(*testing.T, *mimir.Config)) {
+func testSingle(t *testing.T, arguments []string, configYAML string, stdoutMessage, stderrMessage, stdoutExcluded, stderrExcluded string, assertConfig func(*testing.T, *mimir.Config)) {
 	t.Helper()
 	oldArgs, oldStdout, oldStderr, oldTestMode := os.Args, os.Stdout, os.Stderr, testMode
 	restored := false
@@ -336,7 +334,7 @@ func testSingle(t *testing.T, arguments []string, configYAML string, stdoutMessa
 
 	testMode = true
 	os.Args = arguments
-	co := captureOutput(t)
+	co := test.CaptureOutput(t)
 
 	// reset default flags
 	flag.CommandLine = flag.NewFlagSet(arguments[0], flag.ExitOnError)
@@ -347,72 +345,23 @@ func testSingle(t *testing.T, arguments []string, configYAML string, stdoutMessa
 
 	// Restore stdout and stderr before reporting errors to make them visible.
 	restoreIfNeeded()
-	if !bytes.Contains(stdout, stdoutMessage) {
+	if !strings.Contains(stdout, stdoutMessage) {
 		t.Errorf("Expected on stdout: %q, stdout: %s\n", stdoutMessage, stdout)
 	}
-	if !bytes.Contains(stderr, stderrMessage) {
+	if !strings.Contains(stderr, stderrMessage) {
 		t.Errorf("Expected on stderr: %q, stderr: %s\n", stderrMessage, stderr)
 	}
-	if len(stdoutExcluded) > 0 && bytes.Contains(stdout, stdoutExcluded) {
+	if len(stdoutExcluded) > 0 && strings.Contains(stdout, stdoutExcluded) {
 		t.Errorf("Unexpected output on stdout: %q, stdout: %s\n", stdoutExcluded, stdout)
 	}
-	if len(stderrExcluded) > 0 && bytes.Contains(stderr, stderrExcluded) {
+	if len(stderrExcluded) > 0 && strings.Contains(stderr, stderrExcluded) {
 		t.Errorf("Unexpected output on stderr: %q, stderr: %s\n", stderrExcluded, stderr)
 	}
 	if assertConfig != nil {
 		var cfg mimir.Config
-		require.NoError(t, yaml.Unmarshal(stdout, &cfg), "Can't unmarshal stdout as yaml config")
+		require.NoError(t, yaml.Unmarshal([]byte(stdout), &cfg), "Can't unmarshal stdout as yaml config")
 		assertConfig(t, &cfg)
 	}
-}
-
-type capturedOutput struct {
-	stdoutBuf bytes.Buffer
-	stderrBuf bytes.Buffer
-
-	wg                         sync.WaitGroup
-	stdoutReader, stdoutWriter *os.File
-	stderrReader, stderrWriter *os.File
-}
-
-func captureOutput(t *testing.T) *capturedOutput {
-	stdoutR, stdoutW, err := os.Pipe()
-	require.NoError(t, err)
-	os.Stdout = stdoutW
-
-	stderrR, stderrW, err := os.Pipe()
-	require.NoError(t, err)
-	os.Stderr = stderrW
-
-	co := &capturedOutput{
-		stdoutReader: stdoutR,
-		stdoutWriter: stdoutW,
-		stderrReader: stderrR,
-		stderrWriter: stderrW,
-	}
-	co.wg.Add(1)
-	go func() {
-		defer co.wg.Done()
-		io.Copy(&co.stdoutBuf, stdoutR)
-	}()
-
-	co.wg.Add(1)
-	go func() {
-		defer co.wg.Done()
-		io.Copy(&co.stderrBuf, stderrR)
-	}()
-
-	return co
-}
-
-func (co *capturedOutput) Done() (stdout []byte, stderr []byte) {
-	// we need to close writers for readers to stop
-	_ = co.stdoutWriter.Close()
-	_ = co.stderrWriter.Close()
-
-	co.wg.Wait()
-
-	return co.stdoutBuf.Bytes(), co.stderrBuf.Bytes()
 }
 
 func TestExpandEnvironmentVariables(t *testing.T) {